+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project are documented here. The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project uses
 [semantic versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.0
+
+ - Updated IP colors, using yellow on red rather than gray on red for
+   increased readability.
+
 ## 0.4.4
 
 ### Fixed